Refactoring the intervals
