+++ src/interface/interval-declaration.ts
@@ -1,5 +1,5 @@
 import { DocumentUri, Range } from 'vscode-languageserver';
-import { VariableDeclaration } from '../variable/variable-declaration';
+import { VariableDeclaration } from './variable/variable-declaration';
 
 export interface IntervalDeclaration {
     originalRange: Range;
